requirements.txt
pyyaml
requests
openai
ollama
rich
